refactor: Update import paths for EnhancedJsonMapping and add new file for enhanced type validation
